+++ src/Lucid/Relations/BaseRelation.js
@@ -185,7 +185,11 @@ class BaseRelation {
    * @return {Object}
    */
   async eagerLoad (rows) {
-    this._eagerLoadFn(this.relatedQuery, this.foreignKey, this.mapValues(rows))
+    const mappedRows = this.mapValues(rows)
+    if (!mappedRows || !mappedRows.length) {
+      return this.group([])
+    }
+    this._eagerLoadFn(this.relatedQuery, this.foreignKey, mappedRows)
     const relatedInstances = await this.relatedQuery.fetch()
     return this.group(relatedInstances.rows)
   }